Moved functions specific to search results to separate file
